+++ Sources/RAW/Protocols/RAW_comparable.swift
@@ -9,11 +9,11 @@ public protocol RAW_comparable:RAW_convertible {
 
 extension RAW_comparable {
 	/// compare two RAW_comparable types in memory, using the RAW_compare function. this typically requires a temporary byte buffer to be allocated in order for the comparison to be performed.
-	public static func RAW_compare(lhs:Self, rhs:Self) -> Int32 {
-		lhs.RAW_access({ buff, size in
-			rhs.RAW_access({ buff2, size2 in
-				return Self.RAW_compare(lhs_data:buff, lhs_count:size, rhs_data:buff2, rhs_count:size2)
-			})
-		})
-	}
+	// public static func RAW_compare(lhs:Self, rhs:Self) -> Int32 {
+	// 	lhs.RAW_access({ buff, size in
+	// 		rhs.RAW_access({ buff2, size2 in
+	// 			return Self.RAW_compare(lhs_data:buff, lhs_count:size, rhs_data:buff2, rhs_count:size2)
+	// 		})
+	// 	})
+	// }
 }