+++ Sources/RAW_macros/Plugin.swift
@@ -6,6 +6,8 @@ import SwiftSyntax
 @main
 struct RAW_macros:CompilerPlugin {
 	let providingMacros:[Macro.Type] = [
+		RAW_convertible_string_init_macro.self,
+		RAW_convertible_string_type_macro.self,
 		RAW_staticbuff_binaryfloatingpoint_init_macro.self,
 		RAW_staticbuff_floatingpoint_type_macro.self,
 		RAW_staticbuff_fixedwidthinteger_bridge_macro.self,